Recreated imfuse, updated pad_image, included visualizations
piecewiseaffine_render:
* Update pad_image to separate inputs for padding on the low edges
of the image, as well as padding on the high edges.
*  Create imfuse to take two images, their spatial references (still
to be defined as an object, so just a 2-element vector), and merge
into one image.
* Start imfuse_mesh to apply mesh grids to merged images from
imfuse.

visualization:
* Adapt create_color_plots.py from StitchEM repo to Dodam's mesh
style.
* Create visualization.jl to use PyCall to call Python code.

Remove piecewiseaffine_example, because demos in piecewise_render
suffice to show how it works. Update gitignore to not commit
compiled Python code, or large images from demo methods.