Fix imwarp to accommodate offset and unit test
Update imwarp to take 2-element array as offset for input image
to the global origin. The offset is continuous, and is applied
to the image before the transform is applied.

Update BoundingBox methods to accommodate change from 2xN to Nx2
vector sets, and right-handed matrix multiplication when
tranforming bounding box vertices. Adjust matrix multiplication
in imwarp to also be right-handed. i.e.
    [x y 1] * M
    M = [a b 0;
         c d 0;
         i j 1]

Update imwarp comments to explain logic in switching between
pixel space and global space. Remember: the origin in pixel space
starts at (1,1) in global space. Transforms are intended to be
applied to images in the global space. For example, a rotation
applied in the global space will maintain the upper-left pixel
as the point of rotation. But if that same rotation were applied
in the pixel space, the upper-left hand pixel would no longer be
the point of rotation.

Include unit tests for imwarp.

Current behavior to note:
* With an identity transform, imwarp will cause the last row and
column of an image to be filled black, due to interpolation.
* Bounding boxes take in an i & j for an offset, then a *height* &
*width* - not the opposite corner's point.