Include test/runtests.jl to have Travis-CI run unit tests.
Include test/runtests.jl to make include call to all the files in
the test folder. It will run all the unit tests when we push to the
git repo, and whenever we ask it to run them from the website.

For now, include any test scripts in the test folder. Name the file
as test_[filename].jl where filename is the location of the
file in the src folder being tested. I'm not organizing each test
file into functions, so that just an include("[test filename]")
will run the test. Not sure that's the right approach.

Include minor tweaks in other files for rendering and displaying
blockmatch images.