Remove imwarp artifacts. Resize draw_indices. Remove prealign pads.
Alter line in imwarp creating destination array via similar method.
Now use zeros, typed to be the input array type. Should remove the
artifacts we've seen in imwarped renderings.

Adjust size and offset of the text annotations in draw_indices. A
fontsize of 12.0 was too small to see in the prealignment
thumbnails.

Remove rescopeimage from prealignment method. Unnecessary, now that
we've decided to prealign the entire stack, instead of operating in
batches. The implementation was also not working - it was likely
buggy.