Update review and visualize to more completely view blockmatches
Update visualize to draw boxes on ImageView canvases and save as
PNG files. Use box drawing method to identify target location in
the source image for a blockmatching pair, as well as the peak in
the cross correlogram.

Moved sort_dir from render.jl to utilities.jl.

Update params for prealignment.

Started pipeline_failure_log to track the issues we have in
running the whole pipeline, what's their root cause, and how we
fix them. The goal is to use this to understand how to better catch
and automate the fixes for these failure modes.