Update prealignment render offset handling. Update movie review.
Update render_prealignment to write out the offset between images,
instead of from the start of the stack. This was the cause of our
problems with alignment - it happened after we switched to
alignment between montaged images.

Update scan_section_movie to properly loop through all the image
sections of a stack. Include overview image to track where in the
stack the current movie is.