Update render_prealigned to handle montage offsets.
Incorporate montage offsets into render_prealigned. Thumbnails are
generated so that the source image is offset at [0,0], while the
destination image is offset relative to the source (the offset
listed in prealigned_offsets.txt).

Remove method efficiency for caching warped images. Thumbnails are
now relatively in alignment (not perfect, until we update the
offsets to be global).